RW-10929 PR comments
